follow proper process for tpen project delete.
The TPEN project needs to be successfully deleted before deleting the
content type. The project id is extracted from the tpen uri to send
to the tpen delete call, along with the session id from a login call.
In case of tpen delete failure, content type delete is aborted, and
an error message is shown to the user.